Move analyses constant
