+++ web/src/components/AnalyzeData.vue
@@ -1,6 +1,6 @@
 <script>
 import Vue from 'vue';
-import { analyses } from '.';
+import { analyses } from './vis';
 
 export default Vue.extend({
   props: {
