+++ web/src/components/vis/index.js
@@ -1,6 +1,6 @@
-import WilcoxonPlotTile from './vis/WilcoxonPlotTile.vue';
-import AnovaTableTile from './vis/AnovaTableTile.vue';
-import { wilcoxon_zero_methods, wilcoxon_alternatives, plot_types } from '../utils/constants';
+import WilcoxonPlotTile from './WilcoxonPlotTile.vue';
+import AnovaTableTile from './AnovaTableTile.vue';
+import { wilcoxon_zero_methods, wilcoxon_alternatives, plot_types } from '../../utils/constants';
 
 export const analyses = [
   {
